// src/app/models/pages/entidad.ts
export class Entidad{

	idCompania: number;
	id: number;
	nombrePrimero: string;
	nombreSegundo: string;
	apellidoPaterno: string;
	apellidoMaterno: string;
	nombreLegal: string;
	nombreComercial: string;
	tipoDocumento: string;
	numeroDocumento: string;
	webpage: string;
	email: string;
	telefono1: string;
	telefono2: string;
	codigoHabido: string;
  idPais: number;
	estado: number;
}