// src/app/models/pages/RepMaeEntidad.ts
export class RepMaeEntidad{

	id_compania: number;
	des_empresa: string;
	id_entidad: number;
	nombre_primero: string;
	nombre_segundo: string;
	apellido_paterno: string;
	apellido_materno: string;
	nombre_legal: string;
	nombre_comercial: string;
	tipo_documento: string;
	des_tipo_doc: string;
	numero_documento: string;
	webpage: string;
	email: string;
	telefono1: string;
	telefono2: string;
	pais: string;
	estado: string;
	tipo_entidad: string;

}